Unify linear and nonlinear RGB
