+++ examples/readme_examples.rs
@@ -26,7 +26,7 @@ mod manipulation {
     use display_colors;
 
     pub fn run() {
-        let color: Color = Srgb::new(0.8, 0.2, 0.1).into();
+        let color: Color = Srgb::new(0.8, 0.2, 0.1).into_linear().into();
         let lighter = color.lighten(0.1);
         let desaturated = color.desaturate(0.5);
