+++ src/convert.rs
@@ -2,7 +2,7 @@ use num_traits::Float;
 
 use {Alpha, Luma, Xyz, Yxy, Lab, Lch, Hsv, Hwb, Hsl, Color};
 use white_point::{WhitePoint, D65};
-use rgb::{RgbSpace, RgbStandard, LinRgb, Rgb};
+use rgb::{RgbSpace, RgbStandard, Rgb, Lin};
 
 ///FromColor provides conversion between the colors.
 ///
@@ -35,18 +35,18 @@ pub trait FromColor<Wp = D65, T = f32>: Sized
     }
 
     ///Convert from RGB color space
-    fn from_rgb<S: RgbSpace<WhitePoint=Wp>>(inp: LinRgb<S, T>) -> Self {
+    fn from_rgb<S: RgbSpace<WhitePoint=Wp>>(inp: Rgb<Lin<S>, T>) -> Self {
         Self::from_xyz(inp.into_xyz())
     }
 
     ///Convert from HSL color space
     fn from_hsl<S: RgbSpace<WhitePoint=Wp>>(inp: Hsl<S, T>) -> Self {
-        Self::from_rgb(LinRgb::<S, T>::from_hsl(inp))
+        Self::from_rgb(Rgb::<Lin<S>, T>::from_hsl(inp))
     }
 
     ///Convert from HSV color space
     fn from_hsv<S: RgbSpace<WhitePoint=Wp>>(inp: Hsv<S, T>) -> Self {
-        Self::from_rgb(LinRgb::<S, T>::from_hsv(inp))
+        Self::from_rgb(Rgb::<Lin<S>, T>::from_hsv(inp))
     }
 
     ///Convert from HWB color space
@@ -90,19 +90,19 @@ pub trait IntoColor<Wp = D65, T = f32>: Sized
     }
 
     ///Convert into RGB color space.
-    fn into_rgb<S: RgbSpace<WhitePoint=Wp>>(self) -> LinRgb<S, T> {
-        LinRgb::from_xyz(self.into_xyz())
+    fn into_rgb<S: RgbSpace<WhitePoint=Wp>>(self) -> Rgb<Lin<S>, T> {
+        Rgb::from_xyz(self.into_xyz())
     }
 
     ///Convert into HSL color space
     fn into_hsl<S: RgbSpace<WhitePoint=Wp>>(self) -> Hsl<S, T> {
-        let rgb: LinRgb<S, T> = self.into_rgb();
+        let rgb: Rgb<Lin<S>, T> = self.into_rgb();
         Hsl::from_rgb(rgb)
     }
 
     ///Convert into HSV color space
     fn into_hsv<S: RgbSpace<WhitePoint=Wp>>(self) -> Hsv<S, T> {
-        let rgb: LinRgb<S, T> = self.into_rgb();
+        let rgb: Rgb<Lin<S>, T> = self.into_rgb();
         Hsv::from_rgb(rgb)
     }
 
@@ -142,8 +142,8 @@ macro_rules! impl_into_color {
                 Lch::$from_fn(self)
             }
 
-            fn into_rgb<S: RgbSpace<WhitePoint=Wp>>(self) -> LinRgb<S, T> {
-                LinRgb::$from_fn(self)
+            fn into_rgb<S: RgbSpace<WhitePoint=Wp>>(self) -> Rgb<Lin<S>, T> {
+                Rgb::$from_fn(self)
             }
 
             fn into_hsl<S: RgbSpace<WhitePoint=Wp>>(self) -> Hsl<S, T> {
@@ -187,8 +187,8 @@ macro_rules! impl_into_color_rgb {
                 Lch::$from_fn(self)
             }
 
-            fn into_rgb<Sp: RgbSpace<WhitePoint=Wp>>(self) -> LinRgb<Sp, T> {
-                LinRgb::$from_fn(self)
+            fn into_rgb<Sp: RgbSpace<WhitePoint=Wp>>(self) -> Rgb<Lin<Sp>, T> {
+                Rgb::$from_fn(self)
             }
 
             fn into_hsl<Sp: RgbSpace<WhitePoint=Wp>>(self) -> Hsl<Sp, T> {
@@ -218,7 +218,7 @@ macro_rules! impl_from_trait {
             fn from(color: Color<$s, $t>) -> $self_ty {
                 match color {
                     Color::Luma(c) => c.$into_fn(),
-                    Color::LinRgb(c) => c.$into_fn(),
+                    Color::Rgb(c) => c.$into_fn(),
                     Color::Xyz(c) => c.$into_fn(),
                     Color::Yxy(c) => c.$into_fn(),
                     Color::Lab(c) => c.$into_fn(),
@@ -373,22 +373,21 @@ impl_into_color!(Yxy, from_yxy);
 impl_into_color!(Lab, from_lab);
 impl_into_color!(Lch, from_lch);
 impl_into_color!(Luma ,from_luma);
-impl_into_color_rgb!(LinRgb, from_rgb);
 impl_into_color_rgb!(Hsl, from_hsl);
 impl_into_color_rgb!(Hsv, from_hsv);
 impl_into_color_rgb!(Hwb, from_hwb);
 
 
-impl_from_trait!(<S, T> (Xyz<S::WhitePoint, T>, into_xyz) => {LinRgb<S, T>, Hsl<S, T>, Hsv<S, T>, Hwb<S, T>});
-impl_from_trait!(<S, T> (Yxy<S::WhitePoint, T>, into_yxy) => {LinRgb<S, T>, Hsl<S, T>, Hsv<S, T>, Hwb<S, T>});
-impl_from_trait!(<S, T> (Lab<S::WhitePoint, T>, into_lab) => {LinRgb<S, T>, Hsl<S, T>, Hsv<S, T>, Hwb<S, T>});
-impl_from_trait!(<S, T> (Lch<S::WhitePoint, T>, into_lch) => {LinRgb<S, T>, Hsl<S, T>, Hsv<S, T>, Hwb<S, T>});
-impl_from_trait!(<S, T> (Luma<S::WhitePoint, T>, into_luma) => {LinRgb<S, T>, Hsl<S, T>, Hsv<S, T>, Hwb<S, T>});
+impl_from_trait!(<S, T> (Xyz<S::WhitePoint, T>, into_xyz) => {Hsl<S, T>, Hsv<S, T>, Hwb<S, T>});
+impl_from_trait!(<S, T> (Yxy<S::WhitePoint, T>, into_yxy) => {Hsl<S, T>, Hsv<S, T>, Hwb<S, T>});
+impl_from_trait!(<S, T> (Lab<S::WhitePoint, T>, into_lab) => {Hsl<S, T>, Hsv<S, T>, Hwb<S, T>});
+impl_from_trait!(<S, T> (Lch<S::WhitePoint, T>, into_lch) => {Hsl<S, T>, Hsv<S, T>, Hwb<S, T>});
+impl_from_trait!(<S, T> (Luma<S::WhitePoint, T>, into_luma) => {Hsl<S, T>, Hsv<S, T>, Hwb<S, T>});
 
-impl_from_trait!(<S, T> (LinRgb<S, T>, into_rgb) => {Xyz<S::WhitePoint, T>, Yxy<S::WhitePoint, T>, Lab<S::WhitePoint, T>, Lch<S::WhitePoint, T>, Hsl<S, T>, Hsv<S, T>, Hwb<S, T>, Luma<S::WhitePoint, T>});
-impl_from_trait!(<S, T> (Hsl<S, T>, into_hsl) => {Xyz<S::WhitePoint, T>, Yxy<S::WhitePoint, T>, Lab<S::WhitePoint, T>, Lch<S::WhitePoint, T>, LinRgb<S, T>, Hsv<S, T>, Hwb<S, T>, Luma<S::WhitePoint, T>});
-impl_from_trait!(<S, T> (Hsv<S, T>, into_hsv) => {Xyz<S::WhitePoint, T>, Yxy<S::WhitePoint, T>, Lab<S::WhitePoint, T>, Lch<S::WhitePoint, T>, LinRgb<S, T>, Hsl<S, T>, Hwb<S, T>, Luma<S::WhitePoint, T>});
-impl_from_trait!(<S, T> (Hwb<S, T>, into_hwb) => {Xyz<S::WhitePoint, T>, Yxy<S::WhitePoint, T>, Lab<S::WhitePoint, T>, Lch<S::WhitePoint, T>, LinRgb<S, T>, Hsl<S, T>, Hsv<S, T>, Luma<S::WhitePoint, T>});
+impl_from_trait!(<S, T> (Rgb<Lin<S>, T>, into_rgb) => {Xyz<S::WhitePoint, T>, Yxy<S::WhitePoint, T>, Lab<S::WhitePoint, T>, Lch<S::WhitePoint, T>, Hsl<S, T>, Hsv<S, T>, Hwb<S, T>, Luma<S::WhitePoint, T>});
+impl_from_trait!(<S, T> (Hsl<S, T>, into_hsl) => {Xyz<S::WhitePoint, T>, Yxy<S::WhitePoint, T>, Lab<S::WhitePoint, T>, Lch<S::WhitePoint, T>, Hsv<S, T>, Hwb<S, T>, Luma<S::WhitePoint, T>});
+impl_from_trait!(<S, T> (Hsv<S, T>, into_hsv) => {Xyz<S::WhitePoint, T>, Yxy<S::WhitePoint, T>, Lab<S::WhitePoint, T>, Lch<S::WhitePoint, T>, Hsl<S, T>, Hwb<S, T>, Luma<S::WhitePoint, T>});
+impl_from_trait!(<S, T> (Hwb<S, T>, into_hwb) => {Xyz<S::WhitePoint, T>, Yxy<S::WhitePoint, T>, Lab<S::WhitePoint, T>, Lch<S::WhitePoint, T>, Hsl<S, T>, Hsv<S, T>, Luma<S::WhitePoint, T>});
 
 impl_from_trait_other!(<Wp: WhitePoint, T> (Xyz<Wp, T>, into_xyz) => {Yxy<Wp, T>, Lab<Wp, T>, Lch<Wp, T>, Luma<Wp, T>});
 impl_from_trait_other!(<Wp: WhitePoint, T> (Yxy<Wp, T>, into_yxy) => {Xyz<Wp, T>, Lab<Wp, T>, Lch<Wp, T>, Luma<Wp, T>});
@@ -400,19 +399,7 @@ impl_from_trait_other!(<S: RgbStandard, T> (Xyz<<S::Space as RgbSpace>::WhitePoi
 impl_from_trait_other!(<S: RgbStandard, T> (Yxy<<S::Space as RgbSpace>::WhitePoint, T>, into_yxy) => {Rgb<S, T>});
 impl_from_trait_other!(<S: RgbStandard, T> (Lab<<S::Space as RgbSpace>::WhitePoint, T>, into_lab) => {Rgb<S, T>});
 impl_from_trait_other!(<S: RgbStandard, T> (Lch<<S::Space as RgbSpace>::WhitePoint, T>, into_lch) => {Rgb<S, T>});
-impl_from_trait_other!(<S: RgbStandard, T> (LinRgb<S::Space, T>, into_rgb) => {Rgb<S, T>});
 impl_from_trait_other!(<S: RgbStandard, T> (Hsl<S::Space, T>, into_hsl) => {Rgb<S, T>});
 impl_from_trait_other!(<S: RgbStandard, T> (Hsv<S::Space, T>, into_hsv) => {Rgb<S, T>});
 impl_from_trait_other!(<S: RgbStandard, T> (Hwb<S::Space, T>, into_hwb) => {Rgb<S, T>});
 impl_from_trait_other!(<S: RgbStandard, T> (Luma<<S::Space as RgbSpace>::WhitePoint, T>, into_luma) => {Rgb<S, T>});
-impl_from_trait_other!(<S: RgbStandard, T> (Rgb<S, T>, |c| Rgb::from_linear(c.into_rgb())) => {
-    LinRgb<S::Space, T>,
-    Hsl<S::Space, T>,
-    Hsv<S::Space, T>,
-    Hwb<S::Space, T>,
-    Xyz<<S::Space as RgbSpace>::WhitePoint, T>,
-    Yxy<<S::Space as RgbSpace>::WhitePoint, T>,
-    Lab<<S::Space as RgbSpace>::WhitePoint, T>,
-    Lch<<S::Space as RgbSpace>::WhitePoint, T>,
-    Luma<<S::Space as RgbSpace>::WhitePoint, T>
-});
