+++ src/rgb/mod.rs
@@ -7,12 +7,12 @@ use white_point::WhitePoint;
 use pixel::TransferFn;
 use std::any::Any;
 
-pub use self::linear::{LinRgb, LinRgba};
-pub use self::nonlinear::{Rgb, Rgba};
+pub use self::rgb::{Rgb, Rgba};
+pub use self::standards::{Lin, Linear};
 
 pub mod standards;
-mod linear;
-mod nonlinear;
+//mod linear;
+mod rgb;
 
 ///Nonlinear sRGB.
 pub type Srgb<T = f32> = Rgb<standards::Srgb, T>;
@@ -20,9 +20,9 @@ pub type Srgb<T = f32> = Rgb<standards::Srgb, T>;
 pub type Srgba<T = f32> = Rgba<standards::Srgb, T>;
 
 ///Linear sRGB.
-pub type LinSrgb<T = f32> = LinRgb<standards::Srgb, T>;
+pub type LinSrgb<T = f32> = Rgb<Lin, T>;
 ///Linear sRGB with an alpha component.
-pub type LinSrgba<T = f32> = LinRgba<standards::Srgb, T>;
+pub type LinSrgba<T = f32> = Rgba<Lin, T>;
 
 ///An RGB space and a transfer function.
 pub trait RgbStandard {
